Maze solving visualisation
Working with the raw maze is not the best way to solve the final maze. The best way is to generate a graph and solve that using existing shortest path algorithms for graphs. By flood filling dead ends until a junction is released or a start or end point is found. You can take the final maze by evaluating the remaining cells and placing nodes at the junctions and vertexes between them.